+++ tools/psr/cli/main.go
@@ -1,0 +1,8 @@
+// Copyright (c) 2022, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package main
+
+func main() {
+
+}
