+++ tools/psr/workers/spi/worker.go
@@ -1,10 +1,10 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package worker
+package spi
 
 import (
-	"github.com/verrazzano/verrazzano/tools/psr/worker/config"
+	"github.com/verrazzano/verrazzano/tools/psr/workers/config"
 )
 
 type Worker interface {
